♻️ render함수 전체 리팩토링
- 기존의 render함수를 다음과 같이 분리함.
1. 이벤트 위임로직은 실제 DOM에서 다뤄야 하므로render함수의 최상단에 고정
2. 입력받은VNode를 바탕으로 만들어진 가상 DOM을 넣을  DocumentFragment생성.
3. 가상 돔을 만들어주는 로직을 internalRender함수로 분리.
4. 최종적으로 만들어진 가상 돔을 실제 DOM에 한번에 붙임.
