+++ src/core/render/render.ts
@@ -1,7 +1,5 @@
 import type { VNode } from "@/shared/types/vnode";
-import { mapPropToAttr } from "./mapPropToAttr";
-import { attachHandlers } from "./attachHandlers";
-
+import { internalRender } from "./internalRender";
 // (1) 이벤트 위임 초기화 상태 플래그
 let isEventDelegationInitialized = false;
 
@@ -28,39 +26,12 @@ export function render(vnode: VNode, container: Node): void {
   }
   //---------------------------------------------------------
 
-  if (typeof vnode.type === "function") {
-    render(vnode.type(vnode.props), container);
-    return;
-  }
+  // 2) 빌드 단계: DocumentFragment 생성
+  const fragment: DocumentFragment = document.createDocumentFragment();
 
-  //루트노드 생성(연결고리의 시작점) -> 속성 확인하기 Node | HTMLElement
-  const rootNode: HTMLElement = document.createElement(vnode.type);
+  // 3) internalRender로 전체 트리를 fragment에 구축
+  internalRender(vnode, fragment);
 
-  //VNode의 props 순회
-  Object.entries(vnode.props).forEach(([prop, value]) => {
-    if (prop === "children" && value != null && Array.isArray(value)) {
-      value.forEach((child: string | VNode) => {
-        // 1) 문자열 또는 숫자면 텍스트 노드
-        if (typeof child === "string" || typeof child === "number") {
-          rootNode.appendChild(document.createTextNode(String(child)));
-        } else if (child !== null && child !== undefined) {
-          render(child, rootNode);
-        }
-      });
-    } else if (
-      //이벤트 핸들러 props
-      !!prop &&
-      prop !== "children" &&
-      typeof value === "function"
-    ) {
-      const eventHandler = mapPropToAttr(prop);
-      attachHandlers(rootNode, eventHandler, value);
-    }
-    //표준 HTML 속성
-    else if (value != null && value !== false) {
-      const attribute = mapPropToAttr(prop);
-      rootNode.setAttribute(attribute, value);
-    }
-  });
-  container.appendChild(rootNode);
+  // 4) 커밋 단계: fragment를 실제 container에 한 번에 붙임
+  container.appendChild(fragment);
 }
